Encounters attempt to take into account room space.
So we don't have (as many) scenarios with more skeletons than can fit into a room. This is somewhat crude, in that it picks a max_space for each room, even though encounters are placed in rooms at a later point. It also doesn't deal with non-one diameter monsters properly: they don't pack as well, so this is only approximate.